Update on "Support non-tensor inputs and outputs for checkpointed functions."
As mentioned in https://github.com/pytorch/pytorch/issues/52415,
`torch.utils.checkpoint` doesn't support checkpointing for functions which have
non-tensor inputs and outputs.

This PR resolves this issue by ensuring the autograd machinery ignores the
non-tensor inputs and outputs and processes the tensors accordingly.

Differential Revision: [D26507228](https://our.internmc.facebook.com/intern/diff/D26507228/)

#Closes: https://github.com/pytorch/pytorch/issues/52415

[ghstack-poisoned]

+++ torch/csrc/autograd/custom_function.cpp
@@ -102,7 +102,7 @@ variable_list _wrap_outputs(const variable_list &input_vars,
   int num_diff_outputs = 0;
 
 
-  size_t num_tensors = 0;
+  uint32_t num_tensors = 0;
   for (auto i = 0; i < num_outputs; ++i) {
     Variable var = raw_outputs[i];
     if (!var.defined()) {
